refactor: add list component
